pinout: initial version
This is an unfinished version of a helpful pinout diagram for the
MOD-WIFI-ESP8266-DEV module. It still needs some work to look pretty
but may be useful even in this form.

This was created by exporting the Eagle board design as DXF. The
DXF file was cleaned-up using QCad, the rest was done by Inkscape.

Once ready, PNG and PDF versions should be made available too.